+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# 0.2.1
+
+Released 2019-06-26
+
+* Changed `StandardArena`, `SmallArena` and `TinyArena` typedefs to use new `NonzeroGeneration` generation indices for `Option<Index>` size optimization
+* Changed `TinyWrapArena` typedefs to use new `NonzeroWrapGeneration` indices for `Option<Index>` size optimization
+* Added `U64Arena` to replace the old `StandardArena` and `PicoArena` for a `NanoArena`-like type with `NonzeroWrapGeneration`
+* Added typedefs `U64Index`, `StandardIndex`, `SmallIndex`, `TinyIndex`, `NanoIndex` and `PicoIndex` for indices into the corresponding `Arena`s 
+
 # 0.2.0
 
 Released 2019-06-25.